feat(cm): implement clone-all-repos command
This splits out the clone-all-repos to a separate command
to clone all student/group repositories.
