+++ cmd/cm/clone_repo.go
@@ -3,43 +3,35 @@ package main
 import (
 	"bytes"
 	"cmp"
-	"context"
-	"errors"
 	"flag"
 	"fmt"
 	"os"
 	"os/exec"
 	"path/filepath"
-	"slices"
 	"strings"
-	"time"
-
-	"github.com/quickfeed/quickfeed/scm"
-	"go.uber.org/zap"
 )
 
-// cloneRepo clones a repository from GitHub and pulls assignments from the main repository.
+// cloneRepo clones a specified repository (or the current user's repository) from
+// GitHub course's organization. If the -pull flag is provided, it will pull the
+// assignments repository into the user's cloned (and empty) repository.
 //
 // Run with the following command:
 //
-//	cm clone-repo -repo <repo> [-pull] [-all]
+//	cm clone-repo -repo <repo> [-pull]
 //
 // Examples:
 //
 //	cm clone-repo -repo meling-labs
 //	cm clone-repo -repo group-repo
+//	cm clone-repo -repo meling-labs -pull
 //	cm clone-repo -pull
-//	cm clone-repo -all
 //
 // The -repo flag specifies the GitHub repository to clone. If the flag is not set,
 // the repository is assumed to be the GitHub username-labs repository.
 //
 // The -pull flag specifies that assignments should be pulled from the main repository
 // after cloning the repository. This is useful when setting up a new repository after
-// signing up for the course.
-//
-// The -all flag clones all student/group repositories for the current year.
-// The repositories are cloned into the directory <year>/student-repos/<repo>.
+// signing up for the course; the repository should be empty when cloning.
 func cloneRepo(args []string) {
 	if err := loadEnv(); err != nil {
 		exitErr(err, "Error loading environment variables")
@@ -47,18 +39,12 @@ func cloneRepo(args []string) {
 
 	fs := flag.NewFlagSet(cloneRepoCmd, flag.ExitOnError)
 	var repo string
-	var pull, all bool
+	var pull bool
 	fs.StringVar(&repo, "repo", "", "GitHub repository (default: GitHub username-labs)")
 	fs.BoolVar(&pull, "pull", false, "Pull assignments from main repository")
-	fs.BoolVar(&all, "all", false, "Clone all student/group repositories")
 
 	if err := fs.Parse(args); err != nil {
 		exitErr(err, "Error parsing flags")
-	}
-
-	if all {
-		cloneAllRepos()
-		return
 	}
 
 	repo = cmp.Or(repo, githubUsername()+"-labs")
@@ -89,66 +75,6 @@ func cloneRepo(args []string) {
 
 func repoHome() string {
 	return filepath.Join(filepath.Dir(gitRoot), year())
-}
-
-func cloneAllRepos() {
-	ghOrg := courseOrg()
-	ghClient := getSCMClient()
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	// Alternative: use the GitHub CLI to list repositories:
-	// gh repo list dat520-2025 --limit 100 --json name,url --template '{{.name}}'
-	ghRepos, err := ghClient.GetRepositories(ctx, ghOrg)
-	if err != nil {
-		exitErr(err, "Error listing repositories")
-	}
-
-	path := filepath.Join(repoHome(), "student-repos")
-	if err := os.MkdirAll(path, os.ModePerm); err != nil {
-		exitErr(err, "Error creating repository directory")
-	}
-	repoPath := func(repo string) string {
-		return filepath.Join(path, repo)
-	}
-
-	fmt.Printf("Cloning %d student/group repositories into %q\n", len(ghRepos), path)
-
-	for _, scmRepo := range ghRepos {
-		repo := scmRepo.Repo
-		// skipping the main course repository (assignments, info, tests)
-		if slices.Contains(courseRepos, repo) {
-			continue
-		}
-
-		msg := fmt.Sprintf("Cloned %q into %q", repo, path)
-		if exists(repoPath(repo)) {
-			msg = fmt.Sprintf("Repository %q updated", repo)
-		}
-
-		ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
-		_, err := ghClient.Clone(ctx, &scm.CloneOptions{
-			Organization: ghOrg,
-			Repository:   repo,
-			DestDir:      path,
-		})
-		cancel()
-		if err != nil {
-			fmt.Printf("Error cloning %q: %v\n", repo, err)
-			continue
-		}
-		fmt.Println(msg)
-	}
-}
-
-func getSCMClient() scm.SCM {
-	token := os.Getenv("GITHUB_ACCESS_TOKEN")
-	if token == "" {
-		exitErr(errors.New("missing GITHUB_ACCESS_TOKEN"), "cm: GitHub access token required for this operation")
-	}
-	client, err := scm.NewSCMClient(zap.NewNop().Sugar(), token)
-	check(err)
-	return client
 }
 
 // pullAssignments sets up a [repo] remote in repoPath and
